--- RestSerProducts/src/main/java/MainAppRestProd.java
@@ -1,9 +1,0 @@
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MainAppRestProd {
-    public static void main(String[] args) {
-        SpringApplication.run(MainAppRestProd.class, args);
-    }
-}
