--- RestSerProducts/src/main/java/conroller/ProductController.java
@@ -1,38 +1,0 @@
-package conroller;
-
-import lombok.RequiredArgsConstructor;
-import model.ProductDto;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import service.ProductService;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/products")
-public class ProductController {
-    private final ProductService productService;
-
-    @GetMapping
-    public List<ProductDto> findAllProducts() {
-        return productService.findAllProducts();
-    }
-
-    @GetMapping("/{id}")
-    public ProductDto findProductById(@PathVariable Long id) {
-        return productService.findProductById(id).get();
-    }
-
-    @DeleteMapping("/{id}")
-    public void deleteProductById(@PathVariable Long id) {
-        productService.deleteProductById(id);
-    }
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public ProductDto saveNewProduct(ProductDto product) {
-        product.setId(null);
-        return productService.saveNewProduct(product);
-    }
-}
