--- RestSerProducts/src/main/java/model/ProductDto.java
@@ -1,18 +1,0 @@
-package model;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class ProductDto {
-    private Long id;
-    private String title;
-    private Integer price;
-
-    public ProductDto(Product p) {
-        this.id = p.getId();
-        this.title = p.getTitle();
-        this.price = p.getPrice();
-    }
-}
