--- RestSerProducts/src/main/java/repository/ProductRepository.java
@@ -1,7 +1,0 @@
-package repository;
-
-import model.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
