+++ RestSerProducts/src/main/java/ss/spring/cloud/prodocts/model/Product.java
@@ -1,0 +1,35 @@
+package ss.spring.cloud.prodocts.model;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Data
+@Entity
+@NoArgsConstructor
+@Table (name = "products_tbl")
+public class Product {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "product_id")
+    private Long id;
+
+    @Column(name = "title_fld")
+    private String title;
+
+    @Column(name = "price_fld")
+    private int price;
+
+    @Column(name = "created_at")
+    @CreationTimestamp
+    private LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    @UpdateTimestamp
+    private LocalDateTime updatedAt;
+}
