+++ RestSerProducts/src/main/java/ss/spring/cloud/prodocts/model/ProductDto.java
@@ -1,0 +1,18 @@
+package ss.spring.cloud.prodocts.model;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class ProductDto {
+    private Long id;
+    private String title;
+    private Integer price;
+
+    public ProductDto(Product p) {
+        this.id = p.getId();
+        this.title = p.getTitle();
+        this.price = p.getPrice();
+    }
+}
