+++ RestSerProducts/src/main/java/ss/spring/cloud/prodocts/repository/ProductRepository.java
@@ -1,0 +1,9 @@
+package ss.spring.cloud.prodocts.repository;
+
+import ss.spring.cloud.prodocts.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
