+++ RestSerProducts/src/main/resources/application.yaml
@@ -1,0 +1,28 @@
+
+server:
+  port: 12345
+  #servlet:
+    #context-path: /rest
+spring:
+  application:
+    name: rest-ser-products
+  #datasource:
+   # driver-class-name: org.h2.Driver
+   # url: jdbc:h2:mem:mydatabase;MODE=PostgreSQL
+   # username: sa
+   # password:
+  #jpa:
+   # show-sql: true
+   # properties:
+   #   hibernate:
+   #     dialect: org.hibernate.dialect.H2Dialect
+  #h2:
+   # console:
+    #  enabled: true
+    #  settings:
+     #   web-allow-others: false
+eureka:
+  client:
+    service-url: #serviceUrl
+      defaultZone: ${EUREKA_URI:http://localhost:8113/eureka}
+    fetch-registry: true #fetchRegistry or fetch-Registry
