+++ RestSerProducts/src/main/resources/db/migration/V1__init.sql
@@ -13,10 +13,6 @@ VALUES
 ('LaW2', 3700),
 ('LaM1', 6500),
 ('Re2', 4600),
-('StW3', 5900),
-('StM2', 6000),
-('LaW', 6200),
-('StW1', 3500),
-('StW2', 6500),
-('LaM', 3500);
+('StW3', 5900);
+
