+++ client-test/src/main/java/client/test/ClientTestController.java
@@ -1,0 +1,19 @@
+package client.test;
+
+import com.netflix.discovery.EurekaClient;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class ClientTestController {
+    private final EurekaClient eurekaClient;
+
+    @GetMapping
+    public String test() {
+        return "test";
+    }
+
+
+}
