+++ client-test/src/main/java/client/test/EurekaClientApp.java
@@ -1,0 +1,13 @@
+package client.test;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class EurekaClientApp {
+    public static void main(String[] args) {
+        SpringApplication.run(EurekaClientApp.class, args);
+    }
+}
