--- eureka/src/main/java/ss/spring/cloud/EurekaSerApp.java
@@ -1,13 +1,0 @@
-package ss.spring.cloud;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-public class EurekaSerApp {
-    public static void main(String[] args) {
-        SpringApplication.run(EurekaSerApp.class, args);
-    }
-}
