--- eureka/src/main/resources/application.yaml
@@ -1,17 +1,0 @@
-server:
-  port: 8113
-
-eureka:
-  client:
-    registerWithEureka: false
-    fetchRegistry: false
-
-spring:
-  cloud:
-    config:
-      uri: http://localhost:8081
-
-# spring.cloud.config.username=root
-# spring.cloud.config.password=s3cr3t
-# spring.cloud.config.name=conf
-# spring.cloud.config.uri=http://localhost:8081
