+++ ser-eureka/src/main/java/ss/eureka/ser/springcloud/EurekaApp.java
@@ -1,0 +1,13 @@
+package ss.eureka.ser.springcloud;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@SpringBootApplication
+@EnableEurekaServer
+public class EurekaApp {
+    public static void main(String[] args) {
+        SpringApplication.run(EurekaApp.class, args);
+    }
+}
