--- src/main/java/spring/cloud/ss/Hw517Application.java
@@ -1,13 +1,0 @@
-package spring.cloud.ss;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Hw517Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Hw517Application.class, args);
-	}
-
-}
