--- src/test/java/spring/cloud/ss/Hw517ApplicationTests.java
@@ -1,13 +1,0 @@
-package spring.cloud.ss;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Hw517ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
