+++ zool-ser/src/main/java/ss/spring/clo/zuul/ser/ZuulApp.java
@@ -1,0 +1,15 @@
+package ss.spring.clo.zuul.ser;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
+
+@SpringBootApplication
+@EnableEurekaClient
+@EnableZuulProxy
+public class ZuulApp {
+    public static void main(String[] args) {
+        SpringApplication.run(ZuulApp.class, args);
+    }
+}
