+++ zool-ser/src/main/resources/application.yaml
@@ -1,0 +1,20 @@
+spring:
+  application:
+    name: zuul-service-ss
+server:
+  port: 1313
+  servlet:
+    context-path: /cloud
+eureka:
+  client:
+    registerWithEureka: true
+    fetchRegistry: true
+    serviceUrl:
+      defaultZone: ${EUREKA_URI:http://localhost:8113/eureka}
+
+zuul:
+  routes:
+    rest-ser-products: /api/v1/products/**
+    client-test: /test/**
+
+
